Fix fetch repairertypes and bikeTypes
